+++ libs/base_num/src/lib.rs
@@ -14,12 +14,14 @@ mod _internals; // upcasted_op!, impl_ops!
 
 mod int; // Int
 mod float; // Float, FloatConst
+mod quant; // ValueQuant
 
 structural_mods! { // _mods
     _mods {
         pub use super::{
             float::_all::*,
             int::_all::*,
+            quant::_all::*,
         };
     }
 }